pull out search box to component
